refactor(ui): split the Bigrams chart into one file per quadrant
BigramsChart.tsx had grown to 946 lines, past the 750-line split
threshold for UI components, and the analysis work still queued for
this tab would only push it further.

Move each quadrant into its own module and leave BigramsChart.tsx with
the data fetching, the shared aggregates and the grid layout. The
shells reused across quadrants (Quadrant, EmptyQuadrant, LimitSelect,
the interval threshold input, the gram toggle) go to a fifth module so
neither quadrant has to import from a sibling.

Pure move: no test file changes, and the rendered test ids and
translation keys are byte-identical before and after.
