+++ src/renderer/components/analyze/BigramsChart.tsx
@@ -3,17 +3,6 @@
 import { useEffect, useMemo, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import {
-  Bar,
-  BarChart,
-  CartesianGrid,
-  Cell,
-  LabelList,
-  ResponsiveContainer,
-  Tooltip,
-  XAxis,
-  YAxis,
-} from 'recharts'
-import {
   primaryDeviceScope,
   scopeToSelectValue,
   type DeviceScope,
@@ -25,31 +14,20 @@ import type {
   TypingKeymapSnapshot,
 } from '../../../shared/types/typing-analytics'
 import { fetchBigramAggregateForRange } from './analyze-fetch'
-import { bigramPairLabel } from './analyze-bigram-format'
-import {
-  aggregateFingerPairs,
-} from './analyze-bigram-finger'
 import { useKeycodeFingerMap } from './use-keycode-finger-map'
-import {
-  aggregateBigramClasses,
-  CLASSIFIED_CLASSES,
-  type BigramClassAggregate,
-  type BigramClassTotal,
-  type ClassifiedBigramClass,
-} from './analyze-bigram-classes'
-import {
-  avgIkiAtOrAboveThreshold,
-  avgIkiFromHist,
-  percentileFromHist,
-} from './analyze-bigram-heatmap'
-import { BIGRAM_MIN_COUNT } from './analyze-typing-profile'
-import { ALL_PAIRS_LIMIT, EMPTY_STAT_VALUE } from './analyze-constants'
-import { fmtMs } from './analyze-format'
-import { FILTER_SELECT, LIST_LIMIT_OPTIONS } from './analyze-filter-styles'
-import { SegmentedToggle } from './SegmentedToggle'
+import { aggregateBigramClasses } from './analyze-bigram-classes'
+import { ALL_PAIRS_LIMIT } from './analyze-constants'
+import { FILTER_SELECT } from './analyze-filter-styles'
 import type { RangeMs } from './analyze-types'
-import { Stat, TooltipShell } from './analyze-tooltip'
-import { CHART_TICK_FONT_SIZE } from '../../utils/chart-palette'
+import {
+  Quadrant,
+  LimitSelect,
+  PairIntervalThresholdInput,
+  GramToggle,
+} from './bigrams-quadrant-ui'
+import { TopRanking, SlowRanking } from './BigramsRankingTables'
+import { BigramFingerBarChart, type FingerSort } from './BigramsFingerQuadrant'
+import { BigramClassesCoverage, BigramClassesTable } from './BigramsClassesQuadrant'
 
 interface BigramsChartProps {
   uid: string
@@ -81,8 +59,6 @@ interface BigramsChartProps {
   snapshot: TypingKeymapSnapshot | null
   fingerOverrides?: Record<string, FingerType>
 }
-
-type FingerSort = 'desc' | 'asc'
 
 // Stable empty-array reference so the classes aggregate's useMemo dep
 // doesn't churn every render while the Classes quadrant is hidden
@@ -304,643 +280,5 @@ export function BigramsChart({
       </div>
       <div className="min-h-0 flex-1">{body}</div>
     </div>
-  )
-}
-
-const GRAM_OPTIONS: readonly (2 | 3)[] = [2, 3]
-
-const GRAM_LABEL_KEY: Record<2 | 3, string> = {
-  2: 'analyze.bigrams.gramToggle.bigram',
-  3: 'analyze.bigrams.gramToggle.trigram',
-}
-
-interface GramToggleProps {
-  value: 2 | 3
-  onChange: (next: 2 | 3) => void
-}
-
-/** Segmented 2-gram / 3-gram switch — built from the same
- * `SegmentedToggle` primitive as `FilterDimensionToggle` so the Bigrams
- * tab's own toggle reads as the same control family as the rest of the
- * Analyze filter row. */
-function GramToggle({ value, onChange }: GramToggleProps): JSX.Element {
-  const { t } = useTranslation()
-  return (
-    <SegmentedToggle
-      options={GRAM_OPTIONS}
-      value={value}
-      onChange={onChange}
-      labelFor={(option) => t(GRAM_LABEL_KEY[option])}
-      ariaLabel={t('analyze.bigrams.gramToggle.ariaLabel')}
-      testId="analyze-bigrams-gram-toggle"
-    />
-  )
-}
-
-interface QuadrantProps {
-  title: string
-  controls?: React.ReactNode
-  /** Optional single-line notice rendered under the title/controls row
-   * (e.g. the top-N cap warning). Absent by default. */
-  notice?: React.ReactNode
-  children: React.ReactNode
-}
-
-function Quadrant({ title, controls, notice, children }: QuadrantProps): JSX.Element {
-  return (
-    <div className="flex min-h-0 min-w-0 flex-col gap-2 rounded border border-edge p-2">
-      <div className="flex flex-wrap items-center gap-2">
-        <div className="text-xs font-medium text-content">{title}</div>
-        {controls}
-      </div>
-      {notice}
-      <div className="min-h-0 flex-1 overflow-auto pr-1">{children}</div>
-    </div>
-  )
-}
-
-interface LimitSelectProps {
-  value: number
-  onChange: (next: number) => void
-  testId: string
-}
-
-function LimitSelect({ value, onChange, testId }: LimitSelectProps): JSX.Element {
-  const options = LIST_LIMIT_OPTIONS.includes(value)
-    ? LIST_LIMIT_OPTIONS
-    : [...LIST_LIMIT_OPTIONS, value].sort((a, b) => a - b)
-  return (
-    <select
-      value={value}
-      onChange={(e) => onChange(Number(e.target.value))}
-      data-testid={testId}
-      className={FILTER_SELECT}
-    >
-      {options.map((n) => (
-        <option key={n} value={n}>
-          {n}
-        </option>
-      ))}
-    </select>
-  )
-}
-
-interface PairIntervalThresholdInputProps {
-  value: number
-  onChange: (next: number) => void
-  testId: string
-}
-
-/** Compact `[label] [N] [suffix]` control rendered in both fingerIki
- * and slow quadrant headers. The local draft state lets the user blank
- * the field mid-edit without leaking '' upstream — the parent is only
- * notified on blur / Enter, and an empty draft commits as `0`. */
-function PairIntervalThresholdInput({
-  value,
-  onChange,
-  testId,
-}: PairIntervalThresholdInputProps): JSX.Element {
-  const { t } = useTranslation()
-  const [draft, setDraft] = useState<string>(String(value))
-
-  // Sync the draft when the sibling quadrant's input commits a change.
-  useEffect(() => {
-    setDraft(String(value))
-  }, [value])
-
-  const commit = (raw: string): void => {
-    const trimmed = raw.trim()
-    const parsed = trimmed === '' ? 0 : Math.max(0, Math.floor(Number(trimmed)))
-    const next = Number.isFinite(parsed) ? parsed : 0
-    setDraft(String(next))
-    if (next !== value) onChange(next)
-  }
-
-  return (
-    <span className="inline-flex items-center gap-1 text-xs text-content-muted">
-      <span>{t('analyze.bigrams.pairIntervalThreshold.label')}</span>
-      <input
-        type="number"
-        inputMode="numeric"
-        min={0}
-        step={1}
-        value={draft}
-        onChange={(e) => setDraft(e.target.value)}
-        onBlur={(e) => commit(e.target.value)}
-        onKeyDown={(e) => {
-          if (e.key === 'Enter') {
-            e.currentTarget.blur()
-          }
-        }}
-        aria-label={t('analyze.bigrams.pairIntervalThreshold.ariaLabel')}
-        data-testid={testId}
-        className="w-14 rounded border border-edge bg-surface px-1 py-0.5 text-right tabular-nums text-content focus:border-accent focus:outline-none"
-      />
-      <span>{t('analyze.bigrams.pairIntervalThreshold.suffix')}</span>
-    </span>
-  )
-}
-
-type SortKey = 'count' | 'avgIki' | 'sd' | 'p95'
-interface SortState<K extends SortKey> {
-  key: K
-  dir: 'asc' | 'desc'
-}
-
-function compareNumeric(a: number | null, b: number | null, dir: 'asc' | 'desc'): number {
-  if (a === null && b === null) return 0
-  if (a === null) return 1
-  if (b === null) return -1
-  return dir === 'asc' ? a - b : b - a
-}
-
-/** Compares two ranking rows on the currently active sort field. Every
- * sortable field is `number | null`, so a field lookup plus
- * `compareNumeric` replaces a per-field switch for both `TopRanking`
- * and `SlowRanking`. */
-function compareBySortKey<K extends SortKey>(
-  a: Record<K, number | null>,
-  b: Record<K, number | null>,
-  sort: SortState<K>,
-): number {
-  return compareNumeric(a[sort.key], b[sort.key], sort.dir)
-}
-
-/** Toggles direction when the clicked column is already active,
- * otherwise switches to that column defaulting to `desc`. */
-function toggleSort<K extends SortKey>(prev: SortState<K>, key: K): SortState<K> {
-  return prev.key === key
-    ? { key, dir: prev.dir === 'asc' ? 'desc' : 'asc' }
-    : { key, dir: 'desc' }
-}
-
-interface TopRankingProps {
-  entries: readonly TypingBigramTopEntry[]
-  listLimit: number
-  gram: 2 | 3
-}
-
-function TopRanking({ entries, listLimit, gram }: TopRankingProps): JSX.Element {
-  const { t } = useTranslation()
-  const [sort, setSort] = useState<SortState<'count' | 'avgIki' | 'sd'>>({ key: 'count', dir: 'desc' })
-  const toggle = (key: 'count' | 'avgIki' | 'sd'): void => setSort((prev) => toggleSort(prev, key))
-
-  const sliced = useMemo(() => {
-    const arr = [...entries].slice(0, Math.max(listLimit, 0))
-    arr.sort((a, b) => compareBySortKey(a, b, sort))
-    return arr
-  }, [entries, listLimit, sort])
-
-  if (sliced.length === 0) {
-    return <EmptyQuadrant text={t('analyze.bigrams.empty')} />
-  }
-  return (
-    <table className="w-full text-xs" data-testid="analyze-bigrams-top-ranking">
-      <thead className="text-content-muted">
-        <tr>
-          <th className="px-1 py-1 text-right font-medium">#</th>
-          <th className="px-2 py-1 text-left font-medium">{t('analyze.bigrams.column.pair')}</th>
-          <SortHeader
-            align="right"
-            label={t('analyze.bigrams.column.count')}
-            active={sort.key === 'count'}
-            indicator={sortIndicator(sort, 'count')}
-            onClick={() => toggle('count')}
-          />
-          <SortHeader
-            align="right"
-            label={t('analyze.bigrams.column.avgIki')}
-            title={gram === 3 ? t('analyze.bigrams.column.avgIkiTrigramTooltip') : undefined}
-            active={sort.key === 'avgIki'}
-            indicator={sortIndicator(sort, 'avgIki')}
-            onClick={() => toggle('avgIki')}
-          />
-          <SortHeader
-            align="right"
-            label={t('analyze.bigrams.column.sd')}
-            active={sort.key === 'sd'}
-            indicator={sortIndicator(sort, 'sd')}
-            onClick={() => toggle('sd')}
-          />
-        </tr>
-      </thead>
-      <tbody>
-        {sliced.map((entry, i) => (
-          <tr key={entry.ngramId} className="border-t border-surface-dim">
-            <td className="px-1 py-1 text-right tabular-nums text-content-muted">{i + 1}</td>
-            <td className="px-2 py-1 font-mono">{bigramPairLabel(entry.ngramId)}</td>
-            <td className="px-2 py-1 text-right tabular-nums">{entry.count.toLocaleString()}</td>
-            <td className="px-2 py-1 text-right tabular-nums">{fmtMs(entry.avgIki)}</td>
-            <td className="px-2 py-1 text-right tabular-nums">{fmtMs(entry.sd)}</td>
-          </tr>
-        ))}
-      </tbody>
-    </table>
-  )
-}
-
-interface SlowEntry {
-  ngramId: string
-  count: number
-  hist: number[]
-  avgIki: number | null
-  sd: number | null
-  p95: number | null
-}
-
-interface SlowRankingProps {
-  entries: readonly TypingBigramTopEntry[]
-  listLimit: number
-  /** Shared threshold from `pairIntervalThresholdMs` — see
-   * `avgIkiAtOrAboveThreshold` for the bucket-center caveat. */
-  minAvgIkiMs: number
-  gram: 2 | 3
-}
-
-function SlowRanking({ entries, listLimit, minAvgIkiMs, gram }: SlowRankingProps): JSX.Element {
-  const { t } = useTranslation()
-  const [sort, setSort] = useState<SortState<'count' | 'avgIki' | 'sd' | 'p95'>>({ key: 'avgIki', dir: 'desc' })
-  const toggle = (key: 'count' | 'avgIki' | 'sd' | 'p95'): void => setSort((prev) => toggleSort(prev, key))
-
-  const slowEntries = useMemo<SlowEntry[]>(() => {
-    const eligible: SlowEntry[] = []
-    for (const entry of entries) {
-      const avg = avgIkiAtOrAboveThreshold(entry.hist, minAvgIkiMs)
-      if (avg === null) continue
-      eligible.push({
-        ngramId: entry.ngramId,
-        count: entry.count,
-        hist: entry.hist,
-        avgIki: avg,
-        sd: entry.sd,
-        p95: percentileFromHist(entry.hist, 0.95),
-      })
-    }
-    eligible.sort((a, b) => compareBySortKey(a, b, sort))
-    return eligible.slice(0, Math.max(listLimit, 0))
-  }, [entries, listLimit, minAvgIkiMs, sort])
-
-  if (slowEntries.length === 0) {
-    return <EmptyQuadrant text={t('analyze.bigrams.empty')} />
-  }
-  return (
-    <table className="w-full text-xs" data-testid="analyze-bigrams-slow-ranking">
-      <thead className="text-content-muted">
-        <tr>
-          <th className="px-1 py-1 text-right font-medium">#</th>
-          <th className="px-2 py-1 text-left font-medium">{t('analyze.bigrams.column.pair')}</th>
-          <SortHeader
-            align="right"
-            label={t('analyze.bigrams.column.count')}
-            active={sort.key === 'count'}
-            indicator={sortIndicator(sort, 'count')}
-            onClick={() => toggle('count')}
-          />
-          <SortHeader
-            align="right"
-            label={t('analyze.bigrams.column.avgIki')}
-            title={gram === 3 ? t('analyze.bigrams.column.avgIkiTrigramTooltip') : undefined}
-            active={sort.key === 'avgIki'}
-            indicator={sortIndicator(sort, 'avgIki')}
-            onClick={() => toggle('avgIki')}
-          />
-          <SortHeader
-            align="right"
-            label={t('analyze.bigrams.column.sd')}
-            active={sort.key === 'sd'}
-            indicator={sortIndicator(sort, 'sd')}
-            onClick={() => toggle('sd')}
-          />
-          <SortHeader
-            align="right"
-            label={t('analyze.bigrams.column.p95')}
-            active={sort.key === 'p95'}
-            indicator={sortIndicator(sort, 'p95')}
-            onClick={() => toggle('p95')}
-          />
-        </tr>
-      </thead>
-      <tbody>
-        {slowEntries.map((entry, i) => (
-          <tr key={entry.ngramId} className="border-t border-surface-dim">
-            <td className="px-1 py-1 text-right tabular-nums text-content-muted">{i + 1}</td>
-            <td className="px-2 py-1 font-mono">{bigramPairLabel(entry.ngramId)}</td>
-            <td className="px-2 py-1 text-right tabular-nums">{entry.count.toLocaleString()}</td>
-            <td className="px-2 py-1 text-right tabular-nums">{fmtMs(entry.avgIki)}</td>
-            <td className="px-2 py-1 text-right tabular-nums">{fmtMs(entry.sd)}</td>
-            <td className="px-2 py-1 text-right tabular-nums">{fmtMs(entry.p95)}</td>
-          </tr>
-        ))}
-      </tbody>
-    </table>
-  )
-}
-
-function sortIndicator<K extends SortKey>(sort: SortState<K>, key: K): string {
-  if (sort.key !== key) return ''
-  return sort.dir === 'asc' ? ' ▲' : ' ▼'
-}
-
-interface SortHeaderProps {
-  label: string
-  indicator: string
-  align: 'left' | 'right'
-  active: boolean
-  onClick: () => void
-  /** Header tooltip (native `title`). Absent by default — only the
-   * trigram Avg IKI header sets one today. */
-  title?: string
-}
-
-function SortHeader({ label, indicator, align, active, onClick, title }: SortHeaderProps): JSX.Element {
-  return (
-    <th
-      className={`select-none px-2 py-1 font-medium ${align === 'right' ? 'text-right' : 'text-left'}`}
-      title={title}
-    >
-      <button
-        type="button"
-        onClick={onClick}
-        className={`cursor-pointer ${active ? 'text-content' : 'text-content-muted hover:text-content'}`}
-      >
-        {label}
-        {indicator}
-      </button>
-    </th>
-  )
-}
-
-function EmptyQuadrant({ text, testId }: { text: string; testId?: string }): JSX.Element {
-  return (
-    <div className="py-4 text-center text-xs text-content-muted" data-testid={testId}>
-      {text}
-    </div>
-  )
-}
-
-interface FingerBarChartProps {
-  entries: readonly TypingBigramTopEntry[]
-  snapshot: TypingKeymapSnapshot | null
-  fingerOverrides?: Record<string, FingerType>
-  listLimit: number
-  sort: FingerSort
-  /** Shared threshold from `pairIntervalThresholdMs` — see
-   * `avgIkiAtOrAboveThreshold` for the bucket-center caveat. */
-  minAvgIkiMs: number
-}
-
-function BigramFingerBarChart({
-  entries,
-  snapshot,
-  fingerOverrides,
-  listLimit,
-  sort,
-  minAvgIkiMs,
-}: FingerBarChartProps): JSX.Element {
-  const { t } = useTranslation()
-  const fingerMap = useKeycodeFingerMap(snapshot, fingerOverrides)
-  const data = useMemo<BarDatum[]>(() => {
-    if (fingerMap.size === 0) return []
-    const totals = aggregateFingerPairs(entries, fingerMap)
-    const ranked: BarDatum[] = []
-    for (const [pairKey, total] of totals) {
-      const avg = avgIkiAtOrAboveThreshold(total.hist, minAvgIkiMs)
-      if (avg === null) continue
-      const [fromFinger, toFinger] = pairKey.split('_') as [FingerType, FingerType]
-      const fromLabel = t(`analyze.finger.short.${fromFinger}`)
-      const toLabel = t(`analyze.finger.short.${toFinger}`)
-      ranked.push({
-        id: pairKey,
-        label: `${fromLabel} → ${toLabel}`,
-        value: avg,
-        count: total.count,
-        color: fromFinger.startsWith('left-') ? BAR_LEFT : BAR_RIGHT,
-      })
-    }
-    const dir = sort === 'desc' ? 1 : -1
-    ranked.sort((a, b) => dir * (b.value - a.value) || a.id.localeCompare(b.id))
-    return ranked.slice(0, Math.max(listLimit, 0))
-  }, [entries, fingerMap, listLimit, minAvgIkiMs, sort, t])
-
-  if (snapshot === null) {
-    return (
-      <EmptyQuadrant
-        text={t('analyze.bigrams.fingerIki.noSnapshot')}
-        testId="analyze-bigrams-finger-no-snapshot"
-      />
-    )
-  }
-  if (data.length === 0) {
-    return <EmptyQuadrant text={t('analyze.bigrams.empty')} />
-  }
-  return (
-    <div data-testid="analyze-bigrams-finger-bars">
-      <BigramBarChart data={data} yAxisWidth={100} unit="ms" />
-    </div>
-  )
-}
-
-const BAR_LEFT = 'var(--color-accent-hover)'
-const BAR_RIGHT = 'var(--color-danger)'
-
-interface BigramClassesQuadrantProps {
-  /** Classes aggregate computed once by the parent `BigramsChart` and
-   * shared with the sibling quadrant below — see the "1 calculation, 2
-   * consumers" note at the call site. */
-  aggregate: BigramClassAggregate
-  hasSnapshot: boolean
-}
-
-/** Coverage line rendered in the classes quadrant's notice slot —
- * "N% of pairs classified" — so a low value (heavy `unknown` bucket)
- * is visible instead of the 4-class table silently under-representing
- * the period. Hidden while there's no snapshot since the table already
- * explains that state. */
-function BigramClassesCoverage({ aggregate, hasSnapshot }: BigramClassesQuadrantProps): JSX.Element | null {
-  const { t } = useTranslation()
-  if (!hasSnapshot || aggregate.totalCount === 0) return null
-  const percent = Math.round(((aggregate.totalCount - aggregate.unknownCount) / aggregate.totalCount) * 100)
-  return (
-    <div className="text-xs text-content-muted" data-testid="analyze-bigrams-classes-coverage">
-      {t('analyze.bigrams.classes.coverage', { percent })}
-    </div>
-  )
-}
-
-/** Per-class avgIki, `null` (renders "—") whenever the class's sample
- * falls below `BIGRAM_MIN_COUNT` — the same floor the Typing Profile
- * card uses to suppress its Hand balance / SFB labels on thin data. */
-function classAvgOrNull(total: BigramClassTotal): number | null {
-  if (total.count < BIGRAM_MIN_COUNT) return null
-  return avgIkiFromHist(total.hist)
-}
-
-/** Signed `+N ms` / `-N ms` delta, `'—'` when either side is `null`.
- * Positive means alternating hands was faster than staying on the
- * class's hand — the CHI 2018 headline result. */
-function fmtDelta(value: number | null): string {
-  if (value === null) return EMPTY_STAT_VALUE
-  const rounded = Math.round(value)
-  return `${rounded > 0 ? '+' : ''}${rounded} ms`
-}
-
-/** Left / Right / Alternation / Repetition IKI quadrant — the
- * CHI 2018 Table 3 hand-usage classes. Each row's average comes from
- * `avgIkiFromHist` on that class's own folded histogram (never a mean
- * of two other classes' averages), and ΔLeft / ΔRight compare each
- * same-hand class directly against Alternation. */
-function BigramClassesTable({ aggregate, hasSnapshot }: BigramClassesQuadrantProps): JSX.Element {
-  const { t } = useTranslation()
-
-  if (!hasSnapshot) {
-    return (
-      <EmptyQuadrant
-        text={t('analyze.bigrams.classes.noSnapshot')}
-        testId="analyze-bigrams-classes-no-snapshot"
-      />
-    )
-  }
-
-  const avgByClass: Record<ClassifiedBigramClass, number | null> = {
-    left: classAvgOrNull(aggregate.totals.left),
-    right: classAvgOrNull(aggregate.totals.right),
-    alternation: classAvgOrNull(aggregate.totals.alternation),
-    repetition: classAvgOrNull(aggregate.totals.repetition),
-  }
-  const deltaLeft = avgByClass.left !== null && avgByClass.alternation !== null
-    ? avgByClass.left - avgByClass.alternation
-    : null
-  const deltaRight = avgByClass.right !== null && avgByClass.alternation !== null
-    ? avgByClass.right - avgByClass.alternation
-    : null
-
-  return (
-    <div className="flex h-full min-h-0 flex-col gap-2" data-testid="analyze-bigrams-classes-table">
-      <table className="w-full text-xs">
-        <thead className="text-content-muted">
-          <tr>
-            <th className="px-2 py-1 text-left font-medium">{t('analyze.bigrams.column.class')}</th>
-            <th className="px-2 py-1 text-right font-medium">{t('analyze.bigrams.column.avgIki')}</th>
-            <th className="px-2 py-1 text-right font-medium">{t('analyze.bigrams.column.count')}</th>
-          </tr>
-        </thead>
-        <tbody>
-          {CLASSIFIED_CLASSES.map((cls) => (
-            <tr key={cls} className="border-t border-surface-dim">
-              <td className="px-2 py-1">{t(`analyze.bigrams.classes.className.${cls}`)}</td>
-              <td className="px-2 py-1 text-right tabular-nums">{fmtMs(avgByClass[cls])}</td>
-              <td className="px-2 py-1 text-right tabular-nums">{aggregate.totals[cls].count.toLocaleString()}</td>
-            </tr>
-          ))}
-        </tbody>
-      </table>
-      <div
-        className="flex flex-wrap gap-x-3 gap-y-1 text-xs text-content-muted"
-        data-testid="analyze-bigrams-classes-delta"
-        title={t('analyze.bigrams.classes.deltaTooltip')}
-      >
-        <span>{t('analyze.bigrams.classes.deltaLeft')}: {fmtDelta(deltaLeft)}</span>
-        <span>{t('analyze.bigrams.classes.deltaRight')}: {fmtDelta(deltaRight)}</span>
-      </div>
-    </div>
-  )
-}
-
-interface BarDatum {
-  id: string
-  label: string
-  value: number
-  count: number
-  color: string
-}
-
-const BAR_ROW_PX = 24
-const CHART_VERTICAL_PADDING_PX = 16
-
-interface BigramBarChartProps {
-  data: BarDatum[]
-  yAxisWidth: number
-  unit: string
-}
-
-/** Horizontal bar chart shared by the Finger and Key bigram quadrants.
- * Each row is one categorical bar; height is sized to fit the row count
- * so the parent quadrant's `overflow-auto` handles long lists. recharts'
- * native Tooltip provides the cursor-following bubble that matches the
- * Ergonomics tab's bar charts. */
-function BigramBarChart({ data, yAxisWidth, unit }: BigramBarChartProps): JSX.Element {
-  // Floor at 120px so single-row charts don't squeeze the axis labels.
-  const height = Math.max(120, data.length * BAR_ROW_PX + CHART_VERTICAL_PADDING_PX * 2 + 24)
-  return (
-    <div style={{ height }}>
-      <ResponsiveContainer width="100%" height="100%">
-        <BarChart
-          data={data}
-          layout="vertical"
-          margin={{ top: CHART_VERTICAL_PADDING_PX, right: 40, bottom: CHART_VERTICAL_PADDING_PX, left: 4 }}
-        >
-          <CartesianGrid strokeDasharray="3 3" stroke="var(--color-edge)" horizontal={false} />
-          <XAxis
-            type="number"
-            stroke="var(--color-content-muted)"
-            fontSize={CHART_TICK_FONT_SIZE}
-            tickFormatter={(v) => `${Math.round(Number(v))}`}
-          />
-          <YAxis
-            type="category"
-            dataKey="label"
-            stroke="var(--color-content-muted)"
-            fontSize={CHART_TICK_FONT_SIZE}
-            width={yAxisWidth}
-            interval={0}
-          />
-          <Tooltip
-            cursor={{ fill: 'var(--color-surface-dim)' }}
-            content={(p) => <BigramCellTooltip {...p} />}
-          />
-          <Bar dataKey="value" isAnimationActive={false}>
-            {data.map((row) => (
-              <Cell key={row.id} fill={row.color} />
-            ))}
-            <LabelList
-              dataKey="value"
-              position="right"
-              formatter={(v: unknown) => `${Math.round(Number(v))} ${unit}`}
-              style={{ fill: 'var(--color-content-muted)', fontSize: CHART_TICK_FONT_SIZE }}
-            />
-          </Bar>
-        </BarChart>
-      </ResponsiveContainer>
-    </div>
-  )
-}
-
-interface BigramCellTooltipProps {
-  active?: boolean
-  label?: unknown
-  payload?: ReadonlyArray<{ payload?: BarDatum }>
-}
-
-/** recharts content renderer — the default `formatter` path renders a
- * leading separator when the item name is empty, and threading a name
- * through every row would obscure the per-bigram label that's already
- * on the Y axis. Owning the markup keeps the bubble compact. */
-function BigramCellTooltip({ active, label, payload }: BigramCellTooltipProps): JSX.Element | null {
-  const { t } = useTranslation()
-  if (!active || !payload?.length) return null
-  const datum = payload[0]?.payload
-  if (!datum) return null
-  const displayLabel = typeof label === 'string' || typeof label === 'number' ? label : datum.label
-  return (
-    <TooltipShell header={displayLabel}>
-      <Stat
-        label={t('analyze.bigrams.cellTooltipOccurrencesLabel')}
-        value={datum.count.toLocaleString()}
-      />
-      <Stat
-        label={t('analyze.bigrams.cellTooltipAvgIkiLabel')}
-        value={`${Math.round(datum.value)} ms`}
-      />
-    </TooltipShell>
   )
 }
